# HDDS-12756. Speed up TestReconfigShell and TestOzoneDebugShell