## What changes were proposed in this pull request?

Speed up TestReconfigShell and TestOzoneDebugShell by using shared cluster in `NonHATests`.

Remove a small part of `TestReconfigShell` that runs the command with stopped datanode.  (I think testing with `DECOMMISSIONING` node is enough.)

https://issues.apache.org/jira/browse/HDDS-12756

## How was this patch tested?

Before:

```
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 46.81 s -- in org.apache.hadoop.ozone.shell.TestOzoneDebugShell
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 47.26 s -- in org.apache.hadoop.ozone.shell.TestReconfigShell
```

After:

```
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 7.840 s -- in org.apache.ozone.test.NonHATests$OzoneDebugShell
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.076 s -- in org.apache.ozone.test.NonHATests$ReconfigShell
```

https://github.com/adoroszlai/ozone/actions/runs/14226355835